Rewrite main window with notebook
