+++ main.py
@@ -1,7 +1,6 @@
-import tkinter as tk
 from ui.main_window import MainWindow
 
+
 if __name__ == "__main__":
-    root = tk.Tk()
-    app = MainWindow(root)
-    root.mainloop()
+    app = MainWindow()
+    app.mainloop()
