+++ ui/main_window.py
@@ -1,47 +1,26 @@
 import tkinter as tk
 from tkinter import ttk
 
-
-class ComponentTab(ttk.Frame):
-    def __init__(self, parent):
-        super().__init__(parent)
-
-
-class SupplierTab(ttk.Frame):
-    def __init__(self, parent):
-        super().__init__(parent)
+from ui.component_tab import ComponentTab
+from ui.supplier_tab import SupplierTab
+from ui.warehouse_tab import WarehouseTab
+from ui.storekeeper_tab import StorekeeperTab
+from ui.supply_tab import SupplyTab
 
 
-class WarehouseTab(ttk.Frame):
-    def __init__(self, parent):
-        super().__init__(parent)
-
-
-class StorekeeperTab(ttk.Frame):
-    def __init__(self, parent):
-        super().__init__(parent)
-
-
-class SupplyTab(ttk.Frame):
-    def __init__(self, parent):
-        super().__init__(parent)
-
-
-class MainWindow(ttk.Frame):
-    def __init__(self, root: tk.Tk):
-        super().__init__(root)
-        root.title("ІС відділу комплектації")
-        root.geometry("900x600")
-        root.minsize(800, 500)
-        root.rowconfigure(0, weight=1)
-        root.columnconfigure(0, weight=1)
-
-        self.grid(row=0, column=0, sticky="nsew")
+class MainWindow(tk.Tk):
+    def __init__(self):
+        super().__init__()
+        self.title("ІС відділу комплектації")
+        self.geometry("900x600")
+        self.minsize(800, 500)
         self.rowconfigure(0, weight=1)
         self.columnconfigure(0, weight=1)
 
         notebook = ttk.Notebook(self)
         notebook.grid(row=0, column=0, sticky="nsew")
+        notebook.rowconfigure(0, weight=1)
+        notebook.columnconfigure(0, weight=1)
 
         tabs = [
             ("Комплектуючі", ComponentTab(notebook)),
@@ -56,6 +35,5 @@ class MainWindow(ttk.Frame):
 
 
 if __name__ == "__main__":
-    root = tk.Tk()
-    MainWindow(root)
-    root.mainloop()
+    app = MainWindow()
+    app.mainloop()
