+++ ui/warehouse_tab.py
@@ -1,0 +1,6 @@
+from tkinter import ttk
+
+
+class WarehouseTab(ttk.Frame):
+    def __init__(self, parent):
+        super().__init__(parent)
